requirements.txt
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
docker==4.4.0
idna==2.10
iniconfig==1.1.1
packaging==20.7
pluggy==0.13.1
pre-commit==2.9.2
py==1.9.0
PyGithub==1.54.1
pyparsing==2.4.7
pytest==6.1.2
requests==2.25.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
websocket-client==0.57.0